requirements.txt
numpy
seaborn
pandas
tensorflow
keras
Pillow
streamlit
rembg